requirements.txt
Mako==1.0.0
MarkupSafe==0.23
cffi==1.7.0
chardet==2.3.0
colorama==0.3.2
cryptography==1.3.4
get>=0.0.0
html5lib==0.999
mysql-connector-python==1.2.3
paramiko==2.0.2
ply==3.4
post>=0.0.0
public==0.0.37
pyasn1>=0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pygobject==3.14.0
pysftp==0.2.9
python-apt==0.9.3.12
pyxdg==0.25
query-string==0.0.11
request>=0.0.0
requests==2.4.3
setupfiles==0.0.13
six==1.8.0
urllib3==1.9.1
wheel==0.24.0
